+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Added
   - Copy Value At Marker context menu item
   - Tooltips when hovering over viewer signal names
+  - Commands for use in other extensions and API docs
 
 - Fixed
   - Context menu items now appear when right clicking on value display column